🔍 refactor.buttons element 정리
