+++ src/components/cartPage/ItemList.tsx
@@ -2,7 +2,7 @@ import * as t from "./itemList.style";
 import { theme } from "../../style/theme";
 import { useEffect, useState, useCallback, useMemo } from "react";
 import { SquareBadge } from "../../elements/badge/Badge";
-import { MainButton } from "../../elements/Buttons";
+import { MainButton } from "../../elements/buttons/Buttons";
 import OptionModal from "../modal/optionModal/OptionModal";
 import { useAppDispatch, useAppSelector } from "../../redux/store";
 import { setCheckedItems, setOrderData } from "../../redux/reducer/cartSlice";
