+++ src/components/detailPage/productQty/ProductQty.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 
 import * as t from "./productQty.style";
-import { CountButton, OptionCountButton } from "../../../elements/Buttons";
+import { CountButton, OptionCountButton } from "../../../elements/buttons/Buttons";
 import { PropsType } from "./productQty.type";
 import { removeOption } from "../productOptions/optionsHandler";
 import { useTotalPrice } from "./useTotalPrice";
