+++ src/components/modal/optionModal/OptionModal.tsx
@@ -3,7 +3,7 @@ import { theme } from "../../../style/theme";
 import { useEffect, useState } from "react";
 import { openDB } from "idb";
 
-import { MainButton } from "../../../elements/Buttons";
+import { MainButton } from "../../../elements/buttons/Buttons";
 import Modal from "../Modal";
 import { ColorHoverIcon } from "../../../elements/ColorIcons";
 import { useGetOption } from "./useGetOption";
