+++ src/components/modal/searchModal/schForm/searchId/SearchId.tsx
@@ -7,7 +7,7 @@ import { useSearchId } from "./useSearchId";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { SearchIdValidation } from "../../../../../shared/utils/yup";
 import Input from "../../../../../elements/Input";
-import { MainButton } from "../../../../../elements/Buttons";
+import { MainButton } from "../../../../../elements/buttons/Buttons";
 
 type IdValue = {
   userName: string;
