+++ src/components/modal/userInfoModal/UserInfoModal.tsx
@@ -9,7 +9,7 @@ import Modal from "../Modal";
 import Profile from "../../signUpPage/profile/Profile";
 import AddressForm from "../../signUpPage/addressForm/AddressForm";
 import Input from "../../../elements/Input";
-import { MainButton } from "../../../elements/Buttons";
+import { MainButton } from "../../../elements/buttons/Buttons";
 import { useAppSelector } from "../../../redux/store";
 
 interface ModalType {
