+++ src/containers/completePage/completeBody/CompleteBody.tsx
@@ -1,5 +1,5 @@
 import * as t from "./completeBody.style";
-import { MainButton } from "../../../elements/Buttons";
+import { MainButton } from "../../../elements/buttons/Buttons";
 import { useComplete } from "./useCompleteBody";
 import { CompleteType } from "../completeBody/completeBody.type";
 import { useNavigate } from "react-router-dom";
