+++ src/containers/loginPage/LoginForm.tsx
@@ -4,7 +4,7 @@ import { useState, useEffect } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import { openDB } from "idb";
 import { useLoginForm } from "./useLoginForm";
-import { MainButton } from "../../elements/Buttons";
+import { MainButton } from "../../elements/buttons/Buttons";
 import Input from "../../elements/Input";
 import SearchModal from "../../components/modal/searchModal/schModal/SearchModal";
