+++ src/elements/buttons/buttons.type.ts
@@ -1,0 +1,24 @@
+import { Dispatch, SetStateAction } from "react";
+import { OptionListType } from "../../components/detailPage/productOptions/productOptions.type";
+
+export interface BtnType {
+  width?: string;
+  fontSize?: string;
+  fontWeight?: string;
+  color?: string;
+  border?: string;
+  radius?: string;
+  bgColor?: string;
+  padding?: string;
+  hColor?: string;
+  hBorder?: string;
+  hBgColor?: string;
+  children?: React.ReactNode;
+  onClick?: () => void;
+}
+export interface PropsType {
+  id?: number;
+  qty?: number;
+  optionList?: OptionListType;
+  setQty?: Dispatch<SetStateAction<number>>;
+}
